JS Elements overhaul
Proper autoloading, attribute initializer, basically code splitting
